extract scope and frame for reuse
